requirements.txt
opencv-python
numpy
ultralytics
facenet-pytorch
torch
torchvision
mysql-connector-python
pycryptodome
openpyxl